requirements/base.txt

alembic==1.0.10
Click==7.0
Flask==1.0.3
Flask-Admin==1.5.3
Flask-Login==0.4.1
Flask-Migrate==2.5.2
Flask-SQLAlchemy==2.4.0
Flask-Restless-0.17.0 
Flask-WTF==0.14.2
itsdangerous==1.1.0
Jinja2==2.10.1
Mako==1.0.12
MarkupSafe==1.1.1
python-dateutil==2.8.0
python-editor==1.0.4
python-mimeparse-1.6.0
six==1.12.0
mimerender-0.6.0 
SQLAlchemy==1.3.4
Werkzeug==0.15.4
WTForms==2.2.1

